Moved CommandOutput to its own file
