+++ src/main.rs
@@ -1,5 +1,5 @@
-use crate::utils::{check_if_folder_exists, create_folder_if_not_exists};
 use branch::{Branch, BranchType};
+use error::CommandOutput;
 use log::{debug, error, info};
 use parsing::applications_to_string;
 use regex::Regex;
@@ -8,11 +8,12 @@ use std::fs;
 use std::path::PathBuf;
 use std::{error::Error, io::Write};
 use structopt::StructOpt;
-use utils::{run_command_from_list, CommandOutput};
+use utils::{check_if_folder_exists, create_folder_if_not_exists, run_command_from_list};
 mod argo_resource;
 mod argocd;
 mod branch;
 mod diff;
+mod error;
 mod extract;
 mod kind;
 mod minikube;
